requirements.txt
streamlit>=1.40.0
openai>=1.60.0
python-dotenv>=1.0.1
moviepy==1.0.3
imageio-ffmpeg>=0.5.1
Pillow>=10.0.0
numpy>=1.26.0
requests
python-dotenv
openai